Fix auth provider fast refresh lint
